+++ forui/CHANGELOG.md
@@ -36,6 +36,8 @@ This update adds responsive breakpoints, focused outlines & localization! It als
 
 * Add `FFocusedOutline`.
 
+* Add `FDialog.adaptive(...)`.
+
 * **Breaking** Add `focusedOutlineStyle` to `FAccordionStyle` - this only affect users which use the primary constructor.
 
 * **Breaking** Add `focusedOutlineStyle` to `FBottomNavigationBar` - this only affect users which use the primary constructor.